more metrohash is 64 seed changes

+++ Hashes.h
@@ -130,40 +130,40 @@ inline void xxHash64_with_state_test( const void * key, int len, const void *sta
 }
 
 inline void metrohash64_1_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash64_1((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash64_1((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash64_2_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash64_2((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash64_2((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash128_1_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash128_1((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash128_1((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash128_2_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash128_2((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash128_2((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 #if defined(__SSE4_2__) && defined(__x86_64__)
 
 inline void metrohash64crc_1_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash64crc_1((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash64crc_1((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash64crc_2_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash64crc_2((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash64crc_2((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash128crc_1_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash128crc_1((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash128crc_1((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 inline void metrohash128crc_2_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  metrohash128crc_2((const uint8_t *)key,(uint64_t)len,*((uint32_t*)state),(uint8_t *)out);
+  metrohash128crc_2((const uint8_t *)key,(uint64_t)len,*((uint64_t*)state),(uint8_t *)out);
 }
 #endif
 inline void cmetrohash64_1_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  cmetrohash64_1((const uint8_t *)key,(uint64_t)len,*((uint32_t *)state),(uint8_t *)out);
+  cmetrohash64_1((const uint8_t *)key,(uint64_t)len,*((uint64_t *)state),(uint8_t *)out);
 }
 inline void cmetrohash64_1_optshort_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  cmetrohash64_1_optshort((const uint8_t *)key,(uint64_t)len,*((uint32_t *)state),(uint8_t *)out);
+  cmetrohash64_1_optshort((const uint8_t *)key,(uint64_t)len,*((uint64_t *)state),(uint8_t *)out);
 }
 inline void cmetrohash64_2_with_state_test ( const void * key, int len, const void *state, void * out ) {
-  cmetrohash64_2((const uint8_t *)key,(uint64_t)len,*((uint32_t *)state),(uint8_t *)out);
+  cmetrohash64_2((const uint8_t *)key,(uint64_t)len,*((uint64_t *)state),(uint8_t *)out);
 }
 
 inline void fasthash32_test ( const void * key, int len, const void *state, void * out ) {
